@@ -2,9 +2,9 @@
 // versions:
 // 	protoc-gen-go v1.27.1
 // 	protoc        v3.17.1
-// source: v1/hello.proto
+// source: v1/greeter.proto
 
-package hello
+package greeter
 
 import (
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
@@ -29,7 +29,7 @@ type HelloRequest struct {
 func (x *HelloRequest) Reset() {
 	*x = HelloRequest{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_v1_hello_proto_msgTypes[0]
+		mi := &file_v1_greeter_proto_msgTypes[0]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -42,7 +42,7 @@ func (x *HelloRequest) String() string {
 func (*HelloRequest) ProtoMessage() {}
 
 func (x *HelloRequest) ProtoReflect() protoreflect.Message {
-	mi := &file_v1_hello_proto_msgTypes[0]
+	mi := &file_v1_greeter_proto_msgTypes[0]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -55,7 +55,7 @@ func (x *HelloRequest) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use HelloRequest.ProtoReflect.Descriptor instead.
 func (*HelloRequest) Descriptor() ([]byte, []int) {
-	return file_v1_hello_proto_rawDescGZIP(), []int{0}
+	return file_v1_greeter_proto_rawDescGZIP(), []int{0}
 }
 
 type HelloResponse struct {
@@ -63,13 +63,13 @@ type HelloResponse struct {
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
 
-	MyMessage string `protobuf:"bytes,1,opt,name=my_message,json=myMessage,proto3" json:"my_message,omitempty"`
+	Message string `protobuf:"bytes,1,opt,name=message,proto3" json:"message,omitempty"`
 }
 
 func (x *HelloResponse) Reset() {
 	*x = HelloResponse{}
 	if protoimpl.UnsafeEnabled {
-		mi := &file_v1_hello_proto_msgTypes[1]
+		mi := &file_v1_greeter_proto_msgTypes[1]
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		ms.StoreMessageInfo(mi)
 	}
@@ -82,7 +82,7 @@ func (x *HelloResponse) String() string {
 func (*HelloResponse) ProtoMessage() {}
 
 func (x *HelloResponse) ProtoReflect() protoreflect.Message {
-	mi := &file_v1_hello_proto_msgTypes[1]
+	mi := &file_v1_greeter_proto_msgTypes[1]
 	if protoimpl.UnsafeEnabled && x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
@@ -95,51 +95,51 @@ func (x *HelloResponse) ProtoReflect() protoreflect.Message {
 
 // Deprecated: Use HelloResponse.ProtoReflect.Descriptor instead.
 func (*HelloResponse) Descriptor() ([]byte, []int) {
-	return file_v1_hello_proto_rawDescGZIP(), []int{1}
+	return file_v1_greeter_proto_rawDescGZIP(), []int{1}
 }
 
-func (x *HelloResponse) GetMyMessage() string {
+func (x *HelloResponse) GetMessage() string {
 	if x != nil {
-		return x.MyMessage
+		return x.Message
 	}
 	return ""
 }
 
-var File_v1_hello_proto protoreflect.FileDescriptor
+var File_v1_greeter_proto protoreflect.FileDescriptor
 
-var file_v1_hello_proto_rawDesc = []byte{
-	0x0a, 0x0e, 0x76, 0x31, 0x2f, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x12, 0x06, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x22, 0x0e, 0x0a, 0x0c, 0x48, 0x65, 0x6c, 0x6c,
-	0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x2e, 0x0a, 0x0d, 0x48, 0x65, 0x6c, 0x6c,
-	0x6f, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x1d, 0x0a, 0x0a, 0x6d, 0x79, 0x5f,
-	0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x6d,
-	0x79, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x32, 0x41, 0x0a, 0x07, 0x47, 0x72, 0x65, 0x65,
-	0x74, 0x65, 0x72, 0x12, 0x36, 0x0a, 0x05, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x14, 0x2e, 0x61,
-	0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x1a, 0x15, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x48, 0x65, 0x6c, 0x6c,
-	0x6f, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x42, 0x0e, 0x5a, 0x0c, 0x61,
-	0x70, 0x69, 0x2f, 0x76, 0x31, 0x2f, 0x68, 0x65, 0x6c, 0x6c, 0x6f, 0x62, 0x06, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x33,
+var file_v1_greeter_proto_rawDesc = []byte{
+	0x0a, 0x10, 0x76, 0x31, 0x2f, 0x67, 0x72, 0x65, 0x65, 0x74, 0x65, 0x72, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x12, 0x06, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x22, 0x0e, 0x0a, 0x0c, 0x48, 0x65,
+	0x6c, 0x6c, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x22, 0x29, 0x0a, 0x0d, 0x48, 0x65,
+	0x6c, 0x6c, 0x6f, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x18, 0x0a, 0x07, 0x6d,
+	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65,
+	0x73, 0x73, 0x61, 0x67, 0x65, 0x32, 0x41, 0x0a, 0x07, 0x47, 0x72, 0x65, 0x65, 0x74, 0x65, 0x72,
+	0x12, 0x36, 0x0a, 0x05, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x12, 0x14, 0x2e, 0x61, 0x70, 0x69, 0x2e,
+	0x76, 0x31, 0x2e, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
+	0x15, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x52, 0x65,
+	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x42, 0x10, 0x5a, 0x0e, 0x61, 0x70, 0x69, 0x2f,
+	0x76, 0x31, 0x2f, 0x67, 0x72, 0x65, 0x65, 0x74, 0x65, 0x72, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74,
+	0x6f, 0x33,
 }
 
 var (
-	file_v1_hello_proto_rawDescOnce sync.Once
-	file_v1_hello_proto_rawDescData = file_v1_hello_proto_rawDesc
+	file_v1_greeter_proto_rawDescOnce sync.Once
+	file_v1_greeter_proto_rawDescData = file_v1_greeter_proto_rawDesc
 )
 
-func file_v1_hello_proto_rawDescGZIP() []byte {
-	file_v1_hello_proto_rawDescOnce.Do(func() {
-		file_v1_hello_proto_rawDescData = protoimpl.X.CompressGZIP(file_v1_hello_proto_rawDescData)
+func file_v1_greeter_proto_rawDescGZIP() []byte {
+	file_v1_greeter_proto_rawDescOnce.Do(func() {
+		file_v1_greeter_proto_rawDescData = protoimpl.X.CompressGZIP(file_v1_greeter_proto_rawDescData)
 	})
-	return file_v1_hello_proto_rawDescData
+	return file_v1_greeter_proto_rawDescData
 }
 
-var file_v1_hello_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_v1_hello_proto_goTypes = []interface{}{
+var file_v1_greeter_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
+var file_v1_greeter_proto_goTypes = []interface{}{
 	(*HelloRequest)(nil),  // 0: api.v1.HelloRequest
 	(*HelloResponse)(nil), // 1: api.v1.HelloResponse
 }
-var file_v1_hello_proto_depIdxs = []int32{
+var file_v1_greeter_proto_depIdxs = []int32{
 	0, // 0: api.v1.Greeter.Hello:input_type -> api.v1.HelloRequest
 	1, // 1: api.v1.Greeter.Hello:output_type -> api.v1.HelloResponse
 	1, // [1:2] is the sub-list for method output_type
@@ -149,13 +149,13 @@ var file_v1_hello_proto_depIdxs = []int32{
 	0, // [0:0] is the sub-list for field type_name
 }
 
-func init() { file_v1_hello_proto_init() }
-func file_v1_hello_proto_init() {
-	if File_v1_hello_proto != nil {
+func init() { file_v1_greeter_proto_init() }
+func file_v1_greeter_proto_init() {
+	if File_v1_greeter_proto != nil {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_v1_hello_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_v1_greeter_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*HelloRequest); i {
 			case 0:
 				return &v.state
@@ -167,7 +167,7 @@ func file_v1_hello_proto_init() {
 				return nil
 			}
 		}
-		file_v1_hello_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_v1_greeter_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
 			switch v := v.(*HelloResponse); i {
 			case 0:
 				return &v.state
@@ -184,18 +184,18 @@ func file_v1_hello_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_v1_hello_proto_rawDesc,
+			RawDescriptor: file_v1_greeter_proto_rawDesc,
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
-		GoTypes:           file_v1_hello_proto_goTypes,
-		DependencyIndexes: file_v1_hello_proto_depIdxs,
-		MessageInfos:      file_v1_hello_proto_msgTypes,
+		GoTypes:           file_v1_greeter_proto_goTypes,
+		DependencyIndexes: file_v1_greeter_proto_depIdxs,
+		MessageInfos:      file_v1_greeter_proto_msgTypes,
 	}.Build()
-	File_v1_hello_proto = out.File
-	file_v1_hello_proto_rawDesc = nil
-	file_v1_hello_proto_goTypes = nil
-	file_v1_hello_proto_depIdxs = nil
+	File_v1_greeter_proto = out.File
+	file_v1_greeter_proto_rawDesc = nil
+	file_v1_greeter_proto_goTypes = nil
+	file_v1_greeter_proto_depIdxs = nil
 }
