@@ -1,12 +1,12 @@
 // Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
-// source: v1/hello.proto
+// source: v1/greeter.proto
 
 /*
-Package hello is a reverse proxy.
+Package greeter is a reverse proxy.
 
 It translates gRPC into RESTful JSON APIs.
 */
-package hello
+package greeter
 
 import (
 	"context"
