@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 
-package hello
+package greeter
 
 import (
 	context "context"
@@ -95,5 +95,5 @@ var Greeter_ServiceDesc = grpc.ServiceDesc{
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
-	Metadata: "v1/hello.proto",
+	Metadata: "v1/greeter.proto",
 }
